Moved ModelAdminServiceView to admin package.
